Estadisticas BASSE

+++ models/estadistica.js
@@ -5,7 +5,7 @@ module.exports = function(sequelize, Sequalize) {
         sesiones: Sequalize.INTEGER,
         usuarios_analytics: Sequalize.INTEGER,
         duracion_media: Sequalize.INTEGER,
-        rebote: Sequalize.DECIMAL(10,2),
+        rebote: Sequalize.FLOAT,
         nuevas_sesiones: Sequalize.FLOAT,
         usuarios_wp: Sequalize.INTEGER,
         visitas_paginas: Sequalize.INTEGER
